Move update handlers to one package
