+++ internal/server/http/handlers/update/update.go
@@ -1,7 +1,9 @@
 package update
 
 import (
+	"encoding/json"
 	"github.com/go-chi/chi/v5"
+	"github.com/mkolibaba/metrics/internal/common/http/model"
 	"github.com/mkolibaba/metrics/internal/server/http/handlers"
 	"net/http"
 	"strconv"
@@ -35,6 +37,48 @@ func New(updater MetricsUpdater) http.HandlerFunc {
 			updater.UpdateCounter(name, v)
 		default:
 			w.WriteHeader(http.StatusBadRequest)
+		}
+	}
+}
+
+func NewJSON(updater MetricsUpdater) http.HandlerFunc {
+	writeResponse := func(w http.ResponseWriter, t, name string, counter *int64, gauge *float64) {
+		responseBody := model.Metrics{
+			ID:    name,
+			MType: t,
+			Delta: counter,
+			Value: gauge,
+		}
+		if err := json.NewEncoder(w).Encode(responseBody); err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+		}
+	}
+	return func(w http.ResponseWriter, r *http.Request) {
+		if r.Header.Get("Content-Type") != "application/json" {
+			w.WriteHeader(http.StatusUnsupportedMediaType)
+			return
+		}
+		w.Header().Set("Content-Type", "application/json")
+
+		requestBody := &model.Metrics{}
+		if err := json.NewDecoder(r.Body).Decode(requestBody); err != nil {
+			w.WriteHeader(http.StatusBadRequest)
+			return
+		}
+
+		t := requestBody.MType
+		name := requestBody.ID
+
+		switch requestBody.MType {
+		case handlers.MetricGauge:
+			updated := updater.UpdateGauge(requestBody.ID, *requestBody.Value)
+			writeResponse(w, t, name, nil, &updated)
+		case handlers.MetricCounter:
+			updated := updater.UpdateCounter(requestBody.ID, *requestBody.Delta)
+			writeResponse(w, t, name, &updated, nil)
+		default:
+			w.WriteHeader(http.StatusBadRequest)
+			return
 		}
 	}
 }
