+++ internal/server/testutils/testutils.go
@@ -2,6 +2,7 @@ package testutils
 
 import (
 	"encoding/json"
+	"fmt"
 	"maps"
 	"strings"
 	"testing"
@@ -37,6 +38,9 @@ func AssertResponseBody(t *testing.T, want string, got bodyReader) {
 
 func AssertResponseBodyJSON(t *testing.T, want string, got bodyReader) {
 	t.Helper()
+	if len(got.Body()) == 0 && len(want) == 0 {
+		return
+	}
 	gotMap := make(map[string]any)
 	if err := json.Unmarshal(got.Body(), &gotMap); err != nil {
 		t.Errorf("error during parse got: %v", err)
@@ -48,4 +52,12 @@ func AssertResponseBodyJSON(t *testing.T, want string, got bodyReader) {
 	if !maps.Equal(gotMap, wantMap) {
 		t.Errorf("did not get correct response body: want '%s' got '%s'", wantMap, gotMap)
 	}
+}
+
+func CreateGaugeResponseBodyJSON(id string, val float64) string {
+	return fmt.Sprintf("{\"id\": \"%s\", \"type\": \"gauge\", \"value\": %f}", id, val)
+}
+
+func CreateCounterResponseBodyJSON(id string, val int64) string {
+	return fmt.Sprintf("{\"id\": \"%s\", \"type\": \"counter\", \"delta\": %d}", id, val)
 }
